requirements.txt
boto3>=1.26.0
pinecone-client>=2.2.4
litellm>=1.16.8
